Updates to NavBar Changes to be committed: 	modified:   client/src/App.css 	renamed:    client/src/layouts/MainLayout/index.js -> client/src/layouts/MainLayout.js 	deleted:    client/src/layouts/MainLayout/MainNavBar.js 	new file:   client/src/layouts/MainNavBar.js 	renamed:    client/src/layouts/ObjectViewLayout/ObjectNavBar.js -> client/src/layouts/ObjectNavBar.js 	new file:   client/src/layouts/ObjectViewLayout.js 	deleted:    client/src/layouts/ObjectViewLayout/index.js
