+++ client/src/App.css
@@ -70,31 +70,54 @@
 }
 
 
-/* Object Views */
-
-.viewer-root {
+/* Outlet styles */
+.outlet-root {
   display: flex;
   height: 100%;
   overflow: hidden;
   width: 100%;
 }
 
-.viewer-wrapper {
+.outlet-wrapper {
   display: flex;
   flex: 1 1 auto;
   overflow: hidden;
   padding-top: 64;
 }
 
-.viewer-container {
+.outlet-container {
   display: flex;
   flex: 1 1 auto;
   overflow: hidden;
 }
 
-.viewer-content {
+.outlet-content {
   flex: 1 1 auto;
   height: 100%;
   overflow: auto;
   padding-top: 75px;
+}
+
+/* NavBar styles */
+.grow {
+  flex-grow: 1;
+}
+.image-container {
+  max-width: 100%;
+  height: auto;
+  img {
+    width: "2em";
+  }
+}
+.section-desktop {
+  display: none;
+  @media (min-width: 600px) {
+    display: flex;
+  }
+}
+.section-mobile {
+  display: flex;
+  @media (min-width: 600px) {
+    display: none;
+  }
 }
